ensures the association of reviews with the authenticated user and security in operations

+++ src/main/java/com/th/playnmovie/service/ReviewService.java
@@ -6,6 +6,7 @@ import java.util.stream.Collectors;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 
 import com.th.playnmovie.dto.ReviewDto;
@@ -14,6 +15,7 @@ import com.th.playnmovie.mapper.ReviewMapper;
 import com.th.playnmovie.model.Review;
 import com.th.playnmovie.model.TypeEnum;
 import com.th.playnmovie.repository.ReviewRepository;
+import com.th.playnmovie.security.model.User;
 
 @Service
 public class ReviewService {
@@ -37,9 +39,16 @@ public class ReviewService {
     }
 
     public ReviewDto createReview(ReviewDto reviewDto) {
-        logger.debug("Creating new review: itemId={}, userId={}", reviewDto.getItemId(), reviewDto.getUser().getId());
+        Object main = SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+       if(!(main instanceof User)){
+            throw new IllegalArgumentException("Invalid authenticated user");
+        }
+        User authenticatedUser = (User) main;
+
+        logger.debug("Creating new review: itemId={}, userId={}", reviewDto.getItemId(), authenticatedUser.getId());
 
         Review review = ReviewMapper.fromDto(reviewDto);
+        review.setUser(authenticatedUser);
         reviewRepository.save(review);
 
         logger.info("Review created with id={}", review.getId());
@@ -47,6 +56,12 @@ public class ReviewService {
     }
 
     public ReviewDto updateReview(ReviewDto reviewDto) {
+        Object main = SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        if(!(main instanceof User)){
+            throw new IllegalArgumentException("Invalid authenticated user");
+        }
+        User authenticatedUser = (User) main;
+
         logger.debug("Updating review with id={}", reviewDto.getId());
 
         Review review = reviewRepository.findById(reviewDto.getId())
@@ -54,12 +69,14 @@ public class ReviewService {
                 logger.info("Review not found with id={}", reviewDto.getId());
                 return new ReviewNotFoundException();
             });
+        if(!review.getUser().equals(authenticatedUser)){
+            throw new SecurityException("You can only edit your own reviews.");
+        }
 
         review.setComment(reviewDto.getComment());
         review.setItemId(reviewDto.getItemId());
         review.setNotice(reviewDto.getNotice());
         review.setType(reviewDto.getType());
-        review.setUser(reviewDto.getUser());
 
         reviewRepository.save(review);
 
@@ -68,12 +85,22 @@ public class ReviewService {
     }
 
     public void deleteReview(Long id) {
+    	Object main = SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+    	if(!(main instanceof User)) {
+    		 throw new IllegalArgumentException("Invalid authenticated user");
+    	}
+    	User authenticatedUser = (User)main;
+        
         logger.debug("Deleting review with id={}", id);
 
         Review reviewExists = reviewRepository.findById(id).orElseThrow(() -> {
             logger.info("Review not found for deletion with id={}", id);
             return new ReviewNotFoundException();
         });
+        
+        if(!reviewExists.getUser().equals(authenticatedUser)) {
+        	throw new SecurityException("You can only delete your own reviews.");
+        }
 
         reviewRepository.delete(reviewExists);
         logger.info("Review deleted with id={}", id);
